+++ typescript/packages/subsurface-viewer/src/layers/grid3d/privateGrid3dLayer.ts
@@ -22,17 +22,17 @@ import fsShader from "./fragment.fs.glsl";
 import vsLineShader from "./vertex_lines.glsl";
 import fsLineShader from "./fragment_lines.glsl";
 import { localPhongLighting, utilities } from "../shader_modules";
-import type { TGrid3DColoringMode } from "./grid3dLayer";
+import { TGrid3DColoringMode } from "./grid3dLayer";
 
 const DEFAULT_TEXTURE_PARAMETERS = {
-    [GL.TEXTURE_MIN_FILTER]: GL.LINEAR_MIPMAP_LINEAR,
+    [GL.TEXTURE_MIN_FILTER]: GL.LINEAR,
     [GL.TEXTURE_MAG_FILTER]: GL.LINEAR,
     [GL.TEXTURE_WRAP_S]: GL.CLAMP_TO_EDGE,
     [GL.TEXTURE_WRAP_T]: GL.CLAMP_TO_EDGE,
 };
 
 const DISCRETE_TEXTURE_PARAMETERS = {
-    [GL.TEXTURE_MIN_FILTER]: GL.LINEAR_MIPMAP_LINEAR,
+    [GL.TEXTURE_MIN_FILTER]: GL.NEAREST,
     [GL.TEXTURE_MAG_FILTER]: GL.NEAREST,
     [GL.TEXTURE_WRAP_S]: GL.CLAMP_TO_EDGE,
     [GL.TEXTURE_WRAP_T]: GL.CLAMP_TO_EDGE,
@@ -98,6 +98,17 @@ interface IPropertyUniforms {
     colorMapClampColor: Color | undefined | boolean | number[];
     isColorMapClampColorTransparent: boolean;
     isClampColor: boolean;
+    isColoringDiscrete: boolean;
+    colorMapSize: number;
+}
+
+interface IImageData {
+    data: number[] | Uint8Array;
+    count: number;
+    parameters:
+        | typeof DEFAULT_TEXTURE_PARAMETERS
+        | typeof DISCRETE_TEXTURE_PARAMETERS;
+    isColoringDiscrete: boolean;
 }
 
 // This is a private layer used only by the composite Grid3DLayer
@@ -257,29 +268,45 @@ export default class PrivateLayer extends Layer<PrivateLayerProps> {
         };
     }
 
-    private getImageData(): {
-        imageData: number[] | Uint8Array;
-        count: number;
-        parameters:
-            | typeof DEFAULT_TEXTURE_PARAMETERS
-            | typeof DISCRETE_TEXTURE_PARAMETERS;
-    } {
+    private getDefaultImageData(): IImageData {
+        return {
+            data: new Uint8Array([0, 0, 0]),
+            count: 1,
+            parameters: DISCRETE_TEXTURE_PARAMETERS,
+            isColoringDiscrete: true,
+        };
+    }
+
+    private getImageData(): IImageData {
         if (this.props.colorMapFunction instanceof Uint8Array) {
+            const count = this.props.colorMapFunction.length / 3;
+            if (count === 0) {
+                return this.getDefaultImageData();
+            }
+
+            const parameters =
+                this.props.coloringMode === TGrid3DColoringMode.Property
+                    ? DISCRETE_TEXTURE_PARAMETERS
+                    : DEFAULT_TEXTURE_PARAMETERS;
+            const isColoringDiscrete =
+                this.props.coloringMode === TGrid3DColoringMode.Property;
             return {
-                imageData: this.props.colorMapFunction,
-                count: this.props.colorMapFunction.length / 3,
-                parameters: DISCRETE_TEXTURE_PARAMETERS,
+                data: this.props.colorMapFunction,
+                count,
+                parameters,
+                isColoringDiscrete,
             };
         }
-        const imageData = getImageData(
+        const data = getImageData(
             this.props.colorMapName,
             (this.context as DeckGLLayerContext).userData.colorTables,
             this.props.colorMapFunction
         );
         return {
-            imageData,
+            data,
             count: 256,
             parameters: DEFAULT_TEXTURE_PARAMETERS,
+            isColoringDiscrete: false,
         };
     }
 
@@ -317,7 +344,7 @@ export default class PrivateLayer extends Layer<PrivateLayerProps> {
             width: imageData.count,
             height: 1,
             format: GL.RGB,
-            data: imageData.imageData,
+            data: imageData.data,
             parameters: imageData.parameters,
         });
 
@@ -328,6 +355,8 @@ export default class PrivateLayer extends Layer<PrivateLayerProps> {
             colorMapClampColor,
             isColorMapClampColorTransparent,
             isClampColor,
+            isColoringDiscrete: imageData.isColoringDiscrete,
+            colorMapSize: imageData.count,
         };
     }
 }