+++ typescript/packages/subsurface-viewer/src/utils/BoundingBox3D.test.ts
@@ -1,7 +1,7 @@
 import "jest";
 
 import type { BoundingBox3D } from "./BoundingBox3D";
-import { boxCenter, boxUnion } from "./BoundingBox3D";
+import { boxCenter, boxUnion, isEmpty } from "./BoundingBox3D";
 
 describe("Test BoundingBox3D", () => {
     it("boxUnion default box", () => {
@@ -33,5 +33,25 @@ describe("Test BoundingBox3D", () => {
         expect(boxCenter(box1)).toEqual([2.5, 3.5, 4.5]);
         expect(boxCenter(box2)).toEqual([2.5, 3.5, 4.5]);
         expect(boxCenter(box3)).toEqual([3.5, 4.5, 5.5]);
+    });
+
+    it("isEmpty", () => {
+        const box1: BoundingBox3D = [0, 0, 0, 0, 0, 0];
+        const box2: BoundingBox3D = [1, 2, 3, 4, 5, 3];
+        const box3: BoundingBox3D = [
+            Number.POSITIVE_INFINITY,
+            Number.POSITIVE_INFINITY,
+            Number.POSITIVE_INFINITY,
+            -Number.POSITIVE_INFINITY,
+            -Number.POSITIVE_INFINITY,
+            -Number.POSITIVE_INFINITY,
+        ];
+        expect(isEmpty(undefined)).toBe(true);
+        expect(isEmpty(box1)).toBe(true);
+        expect(isEmpty(box2)).toBe(true);
+        expect(isEmpty(box3)).toBe(true);
+
+        const box4: BoundingBox3D = [1, 2, 3, 4, 5, 6];
+        expect(isEmpty(box4)).toBe(false);
     });
 });